requirements.txt
transformers
accelerate
datasets
evaluate
optimum
sentence-transformers
joblib>=1.2.0
scikit-learn>=1.2.1
pyserini
faiss-cpu
xgboost>=2.0.1